Fix tests + refactor
Tests were failing due to an odd circular import situation. Refactoring the
`buildRequest*` functions into a separate file fixed the issue.
